CAVIAR-package {CAVIAR}	R Documentation

Cambial activity and wood formation: data processing, visualisation and analysis using R

Description

Process, analyses and plots data from cambial activity and wood formation monitoring studies

Details

Package: 	 CAVIAR

Type: 	 Package

Version: 	 2.10-0

Date: 	 2017-01-16

License: 	 GPL (>= 2)

LazyLoad: 	 yes

URL: 	 https://www6.nancy.inra.fr/foret-bois-lerfob/Le-personnel/Fiches-profils/Scientifiques/RATHGEBER-Cyrille

CAVIAR is a package that allow processing, visualisation and analysis of data coming from cambial activity and wood formation monitoring studies.
The developed functions require input data with imposed columns names but not the order. That means that the input table may contain a free number of columns, as long as the compulsory columns, identified by their headers, are provided to the function.

Note

Development started: 27 November 2008.
Last modifications: 05 May 2017.

Author(s)

Cyrille Rathgeber - LERFoB, UMR1092 - INRA Nancy (France)
Email: cyrille.rathgeber@inra.fr

References

Rathgeber C.B.K., Longuetaud F., Mothe F., Cuny H. & Le Moguedec G. 2011. Phenology of wood formation: Data processing, analysis and visualisation using R (Package CAVIAR). Dendrochronologia 29: 139-149.

See Also

aggregateRadialFiles, computeBootstrapTest, computeCriticalDates, computeGompertzCellTimings, computeRingCellNb, expandCellCountTable, fitGompertzModels, GRA, overviewCellCountTable, plotCleavelandDotChart, plotWoodFormationCalendar, plotWoodFormationDynamics, plotWoodGrowth, readExcelCountTable, standardiseCellCounts

[Package CAVIAR version 2.10-0 Index]