GRA {CAVIAR}	R Documentation

GRA cell count table

Description

This dataset was obtained from a cambial activity and wood formation monitoring study (Cuny et al. 2012). Data were collected weekly from 2007 to 2009 for 5 silver firs, 5 Norway spruces, and 5 Scots pines grown in the same stand, located in Grandfontaine forest (Donon mountain, North of Vosges), near Nancy (northeast France).
Column "Site" indicates the code of the site.
Column "Year" indicates the code of the year of sampling.
Column "Species" indicates the tree species code (according to ITRDB).
Column "Tree" indicates the tree reference number.
Column "Sample" indicates the sampling campaign reference number.
Column "DY" indicates the date of the sampling in day of year.
Column "RF" indicates the radial file reference number.
Column "CZ" gives the number of cells in the cambial zone.
Column "EZ" gives the number of cells in the enlarging zone.
Column "WZ" gives the number of cells in the maturing zone.
Column "MZ" gives the number of mature cells.
Column "PR" gives a measure (in number of cells, expression(mu) or mm) of the previous tree-ring width.

Usage

GRA

Format

A dataframe containing 12 columns and 4339 rows

Note

Version: 1.2-3.
Development started: 4 May 2009.
Last modifications: 10 January 2017.

Source

Cyrille Rathgeber - LERFoB, INRA, AgroParisTech, F-54000, Nancy, France
Email: cyrille.rathgeber@inra.fr

References

Cuny H.E., Rathgeber C.B.K., Lebourgeois F., Fortin M. & Fournier M. 2012. Life strategies in intra-annual dynamics of wood formation: example of three conifer species in a temperate forest in north-east France. Tree Physiology 32: 612-625.

See Also

aggregateRadialFiles, computeBootstrapTest, computeCriticalDates, computeGompertzCellTimings, computeRingCellNb, fitGompertzModels, expandCellCountTable, overviewCellCountTable, plotCleavelandDotChart, plotWoodFormationCalendar, plotWoodFormationDynamics, plotWoodGrowth, readExcelCountTable, standardiseCellCounts

[Package CAVIAR version 2.10-0 Index]