aggregateRadialFiles {CAVIAR}	R Documentation

aggregateRadialFiles

Description

Aggregates raw radial files data to compute median or mean cell numbers per tree.

Usage

 aggregateRadialFiles(data, stat = "median")

Arguments

data
a data.frame containing cell count data. This data frame must contain 12 imposed variables with compulsory column names: Site, Year, Tree, Species, Sample, DY, RF, CZ, EZ, WZ, MZ, and PR (see GRA).

stat
an optional character string indicating if RF file should be aggregated using "median" (default) or "mean".

Details

This function takes the raw cell count data (CZ, EZ, WZ, MZ, and PR) recorded by radial file and aggregate (using median or mean) them tree by tree / year.
Additionally, it computes the sums of mature and thickening cells (WMZ = WZ + MZ); and mature, thickening, and enlarging cells (EWMZ = EZ + WZ + MZ).

Value

A data frame of at 13 columns containing: the unchanged group variables: Site, Year, Species, Tree, Sample, and DY;
and the aggregated values for variables: CZ, EZ, WZ, MZ, and PR;
along with the cumulative sums WMZ and EWMZ.

Note

Version: 3.1-2.
Development started: 27 November 2008.
Last modifications: 29 September 2016.

Author(s)

Cyrille Rathgeber - LERFoB, UMR1092 - INRA Nancy (France)
Email: cyrille.rathgeber@inra.fr

References

Rathgeber C.B.K., Longuetaud F., Mothe F., Cuny H. & Le Moguedec G. 2011. Phenology of wood formation: Data processing, analysis and visualisation using R (Package CAVIAR). Dendrochronologia 29: 139-149.

See Also

computeBootstrapTest, computeCriticalDates, computeGompertzCellTimings, computeRingCellNb, expandCellCountTable, fitGompertzModels, GRA, overviewCellCountTable, plotCleavelandDotChart, plotWoodFormationCalendar, plotWoodFormationDynamics, plotWoodGrowth, readExcelCountTable, standardiseCellCounts

Examples

 ## Loading the training dataset:
 data(GRA)

 ## Computing the median for Grandfontaine raw data:
 GRA.median <- aggregateRadialFiles(data=GRA, stat="median")

 ## Computing the mean for Grandfontaine raw data:
 GRA.mean <- aggregateRadialFiles(data=GRA, stat="mean")
 
 ## Creating a csv file containing the aggregated dataset
 write.csv2(GRA.median, file = paste("GRA median - ", Sys.Date(), ".csv", sep=""), row.names = FALSE)

[Package CAVIAR version 2.10-0 Index]