computeBootstrapTest {CAVIAR}	R Documentation

computeBootstrapTest

Description

Computes bootstrap tests for comparing two groups of trees.

Usage

 computeBootstrapTest(data.group1, data.group2, ci.group1 = NA, ci.group2 = NA,
                      stat = "median", centring = FALSE, iter = 1000, out = FALSE,
                      plot = FALSE, main = "NULL", var.name = "?", sub = "NULL")

Arguments

data.group1
a vector containing data for the first group

data.group2
a vector containing data for the second group

ci.group1
an optional vector containing standard deviations for the first group

ci.group2
an optional vector containing standard deviations for the second group

stat
the method to be used:
"mean" for mean comparison;
"Student" for mean comparison using Student's t statistic;
"median" for median comparison (default);
"var" for variance comparison;
"disp.mean" for dispersion around the mean comparison;
"disp.median" for dispersion around the median comparison.

centring
an optional logical indicating if data must be centred or not (default=FALSE)

iter
an optional integer indicating the number of iterations for the bootstrap (default=1000)

out
an optional logical indicating if bootstrapped series must be outputted or not (default=FALSE)

plot
an optional logical indicating if a plot must be outputted or not (default=FALSE)

main
an optional character containing the title of the plot

var.name
an optional character labeling the variable to be tested

sub
an optional character containing the plot sub-title

Details

Computes unilateral bootstrap permutation tests for testing equality of mean, median, or variance between two groups of trees.
Several test statistics can be used: mean difference, Studentized mean difference, median difference, logarithm ratio of variance, dispersion around the mean ratio and dispersion around the median ratio.
For test statistics based on the mean, data can be centred.
Moreover this function can also be feed with two additional vectors containing standard deviations around the critical dates in order to take dispersion into account. It can plot (optionplot) an histogram of the bootstrap distribution of the tested statistic along with its observed value.

Value

A list containing the results of the test:

summary
Summary of the test

ASL
Signification level of the test

ts.obs
Observed value of the tested statistic (out=TRUE)

iterations
number of iterations (out=TRUE)

nb.combinations
number of possible combinations (out=TRUE)

ts
bootstrap series (out=TRUE)

Warning

Dispersion is not implemented for tests based on centred statistics (centring=TRUE)!

Note

Version: 5.1-1
Development started: 14 July 2009.
Last modifications: 22 September 2016.

Author(s)

Cyrille Rathgeber - LERFoB, UMR1092 - INRA Nancy (France)
Email: cyrille.rathgeber@inra.fr

References

Rathgeber C.B.K., Longuetaud F., Mothe F., Cuny H. & Le Moguedec G. 2011. Phenology of wood formation: Data processing, analysis and visualisation using R (Package CAVIAR). Dendrochronologia 29: 139-149.

See Also

aggregateRadialFiles, computeCriticalDates, computeGompertzCellTimings, computeRingCellNb, fitGompertzModels, expandCellCountTable, GRA, overviewCellCountTable, plotCleavelandDotChart, plotWoodFormationCalendar, plotWoodFormationDynamics, plotWoodGrowth, readExcelCountTable, standardiseCellCounts

Examples

 ## Loading the example dataset:
 data(GRA)

 ## Computing wood formation critical dates for firs, spruces, and pines in 2009:
 GRA2009 <- GRA[GRA$Year==2009, ]
 
 Fir2009 <- GRA2009[GRA2009$Sp=="ABAL", ]
 Spruce2009 <- GRA2009[GRA2009$Sp=="PCAB", ]
 Pine2009 <- GRA2009[GRA2009$Sp=="PISY", ]
 
 Fir.cd <- computeCriticalDates(Fir2009, plot = TRUE)
 Spruce.cd <- computeCriticalDates(Spruce2009, plot = TRUE)
 Pine.cd <- computeCriticalDates(Pine2009, plot = TRUE)
 
 ## Testing if fir start xylogenesis before spruce
 computeBootstrapTest(Fir.cd$bE, Spruce.cd$bE)
 
 ## Testing if fir starts xylogenesis before spruce with taking dispersion into account
 computeBootstrapTest(Fir.cd$bE, Spruce.cd$bE, Fir.cd$bE.ci, Spruce.cd$bE.ci)
 
 ## Plotting the test
 computeBootstrapTest(Fir.cd$bE, Spruce.cd$bE, Fir.cd$bE.ci, Spruce.cd$bE.ci, plot=TRUE)

[Package CAVIAR version 2.10-0 Index]