computeCriticalDates {CAVIAR}	R Documentation

computeCriticalDates

Description

Computes wood formation critical dates and durations along with associated confidence intervals

Usage

 computeCriticalDates(data, plot=TRUE)

Arguments

data
a data frame containing cell count data. This data frame must contain at least 7 imposed columns with compulsory names: Tree, DY, RF, CZ, EZ, WZ, MZ (see GRA)

plot
an optional logical indicating if a PDF of the verification plots must be printed (default=TRUE)

Details

This function computes wood formation critical dates and durations from a raw data table containing cell counts (for each radial file) for the 4 development phases (cambial, enlargement, wall thickening, and mature) per sampling dates (see GRA).
Five critical dates as well as their associated confidence intervals are computed using logistic regressions glm, the main date corresponds to the 50% probability of the investigated phase being active while the 95% confidence interval limits corresponds to the 2.5 and 97.5% probabilities.
From the five critical dates, the function computes three critical durations along with their associated confidence intervals (assuming independence between dates).
By default (plot=TRUE), the function outputs a PDF of the verification plots entitled ‘Verification plots for computation of critical dates... .pdf’, illustrating the computation of each critical date for a quick visual checking.

Value

A data frame containing the five dates and the three durations along with their associated confidence intervals.
Critical dates are:

bE
begining of Enlargement period;

bW
begining of Wall thickening period;

bM
begining of Mature period;

cE
cessation of Enlargement period;

cW
cessation of Wall thickening period.

Critical durations are:

dE
duration of Enlargement period (dE = cE - bE);

dW
duration of Wall thickening period (dW = cW - bW);

dX
total duration of wood formation period (dX = cW - bE).

Note

Version: 4.2-1.
Development started: 21 October 2009.
Last modifications: 07 Septembre 2016.

Author(s)

Cyrille Rathgeber - LERFoB, UMR1092 - INRA Nancy (France)
Email: cyrille.rathgeber@inra.fr

References

Rathgeber C.B.K., Longuetaud F., Mothe F., Cuny H. & Le Moguedec G. 2011. Phenology of wood formation: Data processing, analysis and visualisation using R (Package CAVIAR). Dendrochronologia 29: 139-149.

Rathgeber C.B.K., Rossi S.& Bontemps J.-D. 2011. Cambial activity related to tree size in a mature silver-fir plantation. Annals of Botany 108: 429-438.

See Also

aggregateRadialFiles, computeBootstrapTest, computeGompertzCellTimings, computeRingCellNb, expandCellCountTable, fitGompertzModels, GRA, overviewCellCountTable, plotCleavelandDotChart, plotWoodFormationCalendar, plotWoodFormationDynamics, plotWoodGrowth, readExcelCountTable, standardiseCellCounts

Examples

 ## Loading the example dataset:
 data(GRA)
 
 ## Selecting only one year because the function is not taking factor year automatically into account!
 GRA2009 <- GRA[GRA$Year==2009, ]
 
 ## Computing critical dates for all the trees of GRA dataset for year 2009
 GRA2009.CD <- computeCriticalDates(GRA2009, plot = TRUE)
 
 ## Creating a csv file containing computed values
 write.csv2(GRA2009.CD, file = paste("Critical dates - GRA2009 - ", Sys.Date(), ".csv", sep=""), row.names = FALSE)

[Package CAVIAR version 2.10-0 Index]