computeGompertzCellTimings {CAVIAR}	R Documentation

computeGompertzCellTimings

Description

Data analysis function for computing individual cell timings from Gompertz parameters.

Usage

 computeGompertzCellTimings(gompar)

Arguments

gompar
a dataframe containing three sets of Gompertz parameters (a, b, k)

Details

This function computes cell timings using the parameters of Gompertz functions fitted on nELM, nLM, and nM.

Value

A data frame of at least 9 columns containing the cell number (N), and the computed values (tE, tL, tM, dE, dL and dT).

Note

Version: 1.1-0.
Development started: 11 June 2012.
Last modifications: 19 June 2012.

Author(s)

Cyrille Rathgeber - LERFoB, UMR1092 - INRA Nancy (France)
Email: cyrille.rathgeber@inra.fr

See Also

aggregateRadialFiles, computeBootstrapTest, computeCriticalDates, computeRingCellNb, fitGompertzModels, expandCellCountTable, GRA, overviewCellCountTable, plotCleavelandDotChart, plotWoodFormationCalendar, plotWoodFormationDynamics, plotWoodGrowth, readExcelCountTable, standardiseCellCounts

Examples

 ## Loading the example dataset:
 #data(AMA2006)

 ## Averaging radial files for dominant firs
 ## Don't work anymore since averageRadialFiles() was replaced by aggregateRadialFiles !!!
 ## Work in progress...
 #DF <- averageRadialFiles(AMA2006[AMA2006$Sp=="ABAL" & AMA2006$Crown=="Dom", ])

 ## Compute the ring cell number to constrain the asymptote of the Gompertz
 #RCN <- computeRingCellNb(DF, stat="median")[, c("Tree", "Year", "RCN.median")]
 #names(RCN) <- c("Tree", "Year", "A")

 ## Fitting Gompertz function for nELM, nLM and nM for the dominant firs using fixed asymptotes
 #GP <- fitGompertz(DF, parameters=RCN, return.pred=FALSE, plot.fitting=FALSE, plot.OP=FALSE)

 ## Plotting Gompertz function for nELM, nLM and nM for tree 1
 #GP1 <- GP[GP$Tree==1, ]
 #GCT1 <- computeGompertzCellTimings(GP1)

[Package CAVIAR version 2.10-0 Index]