expandCellCountTable {CAVIAR}	R Documentation

expandCellCountTable

Description

Expands a cell count table so that all the missing values are coded explicitly with NA.

Usage

   expandCellCountTable(data)

Arguments

data
a data.frame containing cell count data. This data frame must contain 12 imposed variables with compulsory column names: Site, Year, Tree, Species, Sample, DY, RF, C, E, W, M, and P (see GRA).

Details

This function takes a data frame of raw cell count data (C, E, W, M, and P) recorded by radial file or the aggregated values (median or mean) per tree by tree as input and expand it, so that all the missing values are coded explicitly with NA.

Value

A data frame containing:
the unchanged group variables: Site, Year, Species, Tree, Sample, and DY (plus RF according to input);
and the values for variables: C, E, W, M, and P;

Note

Version: 1.1-0.
Development started: 27 May 2014.
Last modifications: 28 July 2014.

Author(s)

Cyrille Rathgeber - LERFoB, UMR1092 - INRA Nancy (France)
Email: cyrille.rathgeber@inra.fr

References

Rathgeber C.B.K. et al. Wood formation monitoring data verification using CaviaR R package. in prep.

See Also

aggregateRadialFiles, computeBootstrapTest, computeCriticalDates, computeGompertzCellTimings, computeRingCellNb, fitGompertzModels, GRA, overviewCellCountTable, plotCleavelandDotChart, plotWoodFormationCalendar, plotWoodFormationDynamics, plotWoodGrowth, readExcelCountTable, standardiseCellCounts

Examples

   ## Loading the example dataset
   data(GRA)
   
   ## Computing the expanded data frame for GRA raw data
   GRA.exp <- expandCellCountTable(GRA)
   
   ## Computing the expanded data frame for GRA aggregated data
   GRA.med <- aggregateRadialFiles(GRA)
   GRA.med.exp <- expandCellCountTable(GRA.med)

[Package CAVIAR version 2.10-0 Index]