fitGompertzModels {CAVIAR}	R Documentation

fitGompertzModels

Description

fits Gompertz models to the observations.

Usage

 fitGompertzModels(data, parameters = NULL, predictions = FALSE, plot.fitting = FALSE, plot.obs.vs.pre = FALSE)

Arguments

data
a data frame containing cell count data. This data frame must contain at least 4 columns for grouping variables with compulsory names (Site, Year, Tree, and Species),
and at least two columns, but potentially four, with compulsory names for date (DY), and cell count data (EWMZ, WMZ and MZ).

parameters
an optional data frame containing the grouping variable (Site, Year, Tree, and Species) and the value of the asymptotes (A - the same value will be applied for EWMZ, WMZ and MZ);
or the values of all the Gompertz parameters (a, b, k).

predictions
an optional bolean indicating if predicted values must be returned or not (default).

plot.fitting
an optional bolean indicating if a graph of the fitting must be plotted or not (default).

plot.obs.vs.pre
an optional bolean indicating if a verification graph must be plotted or not (default).

Details

This function fits a Gompertz model to the observations and outputs the three fitted parameters, along with six biologically meaningful parameters, and three estimators of the goodness-of-fit.
The function can also outputs the predicted values.
And additionally it draws fitting and verification plots if required.
The Gompertz model is written:

N(t) = A e^{-e^{β - κ t}}

with N(t) the cumulative number of cells and t the day of year.

Value

By default a data frame containing the three fitted parameters, the six biologically meaningful parameters, and the three estimators of the goodness-of-fit :

a
parameter A of the model;

b
parameter β of the model;

k
parameter κ of the model;

bG
date at which 5 % of the cells are produced;

cG
date at which 95 % of the cells are produced;

dG
time elapsed between bG and cG;

tGx
date at which the cell production is maximal (Gompertz curve inflection point);

rGx
maximal rate of cell production;

rGa
average rate of cell production (computed between bG and cG);

R2
R-square;

EF
Modelling Efficiency;

RMSD
Root Mean Squared Deviation.

If option predictions = TRUE, a list of three data frames containing the references of the group (references),
the fitting coefficients (coefficients), and the values of observations and model predictions (values) is returned.

Note

Version: 5.2-4.
Development started: 27 November 2008.
Last modifications: 11 November 2016.

Author(s)

Cyrille Rathgeber - LERFoB, UMR1092 - INRA Nancy (France)
Email: cyrille.rathgeber@inra.fr

References

Rathgeber C.B.K., Rossi S.& Bontemps J.-D. 2011. Cambial activity related to tree size in a mature silver-fir plantation. Annals of Botany 108: 429-438.

See Also

aggregateRadialFiles, computeBootstrapTest, computeCriticalDates, computeGompertzCellTimings, computeRingCellNb, expandCellCountTable, GRA, overviewCellCountTable, plotCleavelandDotChart, plotWoodFormationCalendar, plotWoodFormationDynamics, plotWoodGrowth, readExcelCountTable, standardiseCellCounts

Examples

 ## Loading the training dataset
 data(GRA)
 
 ## Computing the mean for Grandfontaine raw data
 GRA.mean <- aggregateRadialFiles(data=GRA, stat="mean")

 ## Standardising data aggregated using the mean
 GRA.sdd <- standardiseCellCounts(GRA.mean)

 ## Computing the final total ring cell number to constrain the asymptote of the Gompertz
 GRA.sdd_RCN <- computeRingCellNb(GRA.sdd)[, c("Site", "Year", "Tree", "Species", "RCN.Median")]
 names(GRA.sdd_RCN) <- c("Site", "Year", "Tree", "Species", "A")
 
 ## Fitting Gompertz models on standardised data using fixed assymptotes, and printing verification plots in a pdf
 Plot.file.name <- paste("Verification plots for the fitting of Gompertz functions - ",
                         Sys.Date(), ".pdf", sep="")
 pdf(file = Plot.file.name)
 GRA.sdd_GZP.fix <- fitGompertzModels(GRA.sdd, parameters=GRA.sdd_RCN, predictions=FALSE,
                                      plot.fitting=TRUE, plot.obs.vs.pre=TRUE)
 dev.off()

[Package CAVIAR version 2.10-0 Index]