overviewCellCountTable {CAVIAR}	R Documentation

overviewCellCountTable

Description

Check the format and provides a short summary of cell count data frame

Usage

   overviewCellCountTable(data)

Arguments

data
a data.frame containing cell count data.
This data.frame must contain 12 columns with the qualitative variables: Site, Year, Species, Tree, Sample, DY, RF, and the quantitive variables CZ, EZ, WZ, MZ, and PR. (see GRA).

Details

This function is designed to allow a first basic general cheking of cell count data, and print a short summary of factors and quantitative variables contained in the data frame.

Value

None
Information gathered by this function are directly printed in the console.

Note

Version: 1.1-1.
Development started: 22 January 2016.
Last modifications: 10 January 2017.

Author(s)

Cyrille Rathgeber - LERFoB, INRA, AgroParisTech, F-54000, Nancy, France
Email: cyrille.rathgeber@inra.fr

References

None

See Also

aggregateRadialFiles, computeBootstrapTest, computeCriticalDates, computeGompertzCellTimings, computeRingCellNb, fitGompertzModels, expandCellCountTable, GRA, plotCleavelandDotChart, plotWoodFormationCalendar, plotWoodFormationDynamics, plotWoodGrowth, readExcelCountTable, standardiseCellCounts

Examples

   ## Loading the GRA dataset:
   data(GRA)
   
   ## Overview of GRA dataset
   ## -----------------------
   overviewCellCountTable(GRA)

[Package CAVIAR version 2.10-0 Index]