plotCleavelandDotChart {CAVIAR}	R Documentation

plotCleavelandDotChart

Description

Plots Cleaveland Dot Chart for cell count data

Usage

 plotCleavelandDotChart(data, show.previous.ring = TRUE)

Arguments

data
a data frame containing cell count data with imposed column names: Site, Year, Tree, Species, Sample, DY, RF, CZ , EZ, WZ, MZ, and optionally PR

show.previous.ring
an optional logical indicating if previous ring measurements must be used (default=TRUE) or not (FALSE)

Details

Cleaveland Dot Chart are plotted for cell count data for each phase of development, and for a whole data set, groupping trees per site and year automatically.

Value

A List of matrix containing lower and upper fences for each groups of trees.

Note

Version: 1.1-4.
Development started: 27 March 2014.
Last modifications: 27 June 2016.

Author(s)

Cyrille Rathgeber - LERFoB, UMR1092 - INRA Nancy (France)
Email: cyrille.rathgeber@inra.fr

References

Rathgeber C.B.K. Wood formation monitoring data verification using CaviaR R package.

See Also

aggregateRadialFiles, computeBootstrapTest, computeCriticalDates, computeGompertzCellTimings, computeRingCellNb, fitGompertzModels, expandCellCountTable, GRA, overviewCellCountTable, plotWoodFormationCalendar, plotWoodFormationDynamics, plotWoodGrowth, readExcelCountTable, standardiseCellCounts

Examples

   ## Loading the GRA dataset:
   data(GRA)

   ## Looking for outliers and errors in GRA dataset
   plotCleavelandDotChart(GRA, show.previous.ring = TRUE) 

[Package CAVIAR version 2.10-0 Index]