readExcelCountTable {CAVIAR}	R Documentation

readExcelCountTable

Description

Reads Excel file containing cell count data in spreadsheets and write the data into a database-like dataframe.

Usage

 readExcelCountTable(excel.file.name,
                     tree.description.sheet.name = "Tree description",
                     sampling.description.sheet.name = "Sampling description")

Arguments

excel.file.name
a character string containing the path to the Excel file.

tree.description.sheet.name
an optional character string containing the name of the sheet to be read for tree description ("Tree description" by default).

sampling.description.sheet.name
an optional character string containing the name of the sheet to be read for sampling description ("Sampling description" by default).

Details

This function reads Excel files containing metadata and cell count data in different spreadsheets and write the data into a database-like dataframe.
It first reads the tree description and the sampling description sheets of the indicated Excel file before to read automatically the different cell count data sheets based on the information given in the tree description file.
The function requires that:

the Excel file contains a tree description sheet, a sampling description sheet, and several cell count data sheets;

cell count data sheets are called using the number of the tree it described, preceded by the word Tree (example: Tree 56); This number must be given in the tree description sheet;

the tree description sheet contains 3 compulsory columns nammed: Tree, Site, and Species; other columns are discarded;

the sampling description sheet contains 3 compulsory columns nammed: Sample, Year, and DY; other columns are discarded;

a cell count data sheet must contain 16 compulsory columns, nammed: Sample, C1, C2, C3, E1, E2, E3, W1, W2, W3, M1, M2, M3, P1, P2, P3; other columns are discarded.

Value

A dataframe of 12 columns containing the qualitative variables: Site, Year, Species, Tree, Sample, DY, RF, and the quantitive variables CZ, EZ, WZ, MZ, and PR.

Warning

This function is based on the read.xls function, so it needs the gdata package to run!
... The gdata package needs Perl to run.
... You may also need to install the gtools package to run the example.

Note

Version: 2.1-3.
Development started: 13 October 2010.
Last modifications: 10 January 2017.

Author(s)

Cyrille Rathgeber - LERFoB, INRA, AgroParisTech, F-54000, Nancy, France
Email: cyrille.rathgeber@inra.fr

References

Rathgeber C.B.K., Longuetaud F., Mothe F., Cuny H. & Le Moguedec G. 2011. Phenology of wood formation: Data processing, analysis and visualisation using R (Package CAVIAR). Dendrochronologia 29: 139-149.

See Also

aggregateRadialFiles, computeBootstrapTest, computeCriticalDates, computeGompertzCellTimings, computeRingCellNb, fitGompertzModels, expandCellCountTable, GRA, overviewCellCountTable, plotCleavelandDotChart, plotWoodFormationCalendar, plotWoodFormationDynamics, plotWoodGrowth, standardiseCellCounts

Examples

 ## Setting the parameters of the function
 #library(gdata)

 ## Defining path names of the Excel files to be read
 GRA2007.excel.file.name <- file.path(path.package('CAVIAR'), 'extdata', 'GRA2007.xlsx')
 GRA2008.excel.file.name <- file.path(path.package('CAVIAR'), 'extdata', 'GRA2008.xlsx')
 GRA2009.excel.file.name <- file.path(path.package('CAVIAR'), 'extdata', 'GRA2009.xlsx')

 ## Reading the Excel files
 GRA2007.raw.data <- readExcelCountTable(excel.file.name = GRA2007.excel.file.name)
 GRA2008.raw.data <- readExcelCountTable(excel.file.name = GRA2008.excel.file.name)
 GRA2009.raw.data <- readExcelCountTable(excel.file.name = GRA2009.excel.file.name)

 ## Building the data frame containing the 3 years for this site
 GRA.raw.data <- rbind(GRA2007.raw.data, GRA2008.raw.data, GRA2009.raw.data)

 ## Writing GRA data frame in an external csv file
 write.csv2(GRA.raw.data, file="GRA.raw.data.csv", row.names=FALSE)

[Package CAVIAR version 2.10-0 Index]