standardiseCellCounts {CAVIAR}	R Documentation

standardiseCellCounts

Description

Standardise aggregated cell count values using previous ring cell count or measurements.

Usage

 standardiseCellCounts(data, na.rm = "FALSE")

Arguments

data
a data frame composed of 13 columns containing the group variables (Site, Year, Species, Tree, Sample, and DY; as well as the quantitative variables (CZ, EZ, WZ, MZ, PR, WMZ, and EWMZ. Typically, this data frame come from aggregateRadialFiles function.

na.rm
logical indicating if lines where PR is missing should be left as they are without applying any standardisation (option "FALSE" by default), or should be removed from the record (option "TRUE").

Details

This function standardises aggregated cell counts (CZ, EZ, WZ, MZ, WMZ, EWMZ) using previous ring cell counts or measurements (PR). ! Insert formula here !

Value

A data frame similar to the input data frame but containing standardised values for the columns: CZ, EZ, WZ, MZ, WMZ, and EWMZ

Note

Version: 2.1-1.
Development started: 25 January 2011.
Last modifications: 27 Septembre 2016.

Author(s)

Cyrille Rathgeber - LERFoB, UMR1092 - INRA Nancy (France)
Email: cyrille.rathgeber@inra.fr

References

Rathgeber C.B.K., Rossi S.& Bontemps J.-D. 2011. Cambial activity related to tree size in a mature silver-fir plantation. Annals of Botany 108: 429-438.

See Also

aggregateRadialFiles, computeBootstrapTest, computeCriticalDates, computeGompertzCellTimings, computeRingCellNb, fitGompertzModels, expandCellCountTable, GRA, overviewCellCountTable, plotCleavelandDotChart, plotWoodFormationCalendar, plotWoodFormationDynamics, plotWoodGrowth, readExcelCountTable

Examples

 ## Loading the training dataset
 data(GRA)
 
 ## Computing the median for Grandfontaine raw data
 GRA.med <- aggregateRadialFiles(data=GRA, stat="median")

 ## Standardising aggregated data, with keeping in data that cannot be standardised
 GRA.med.sdd1 <- standardiseCellCounts(GRA.med)

 ## Standardising aggregated data, and removing data that can not standardised
 GRA.med.sdd2 <- standardiseCellCounts(GRA.med, na.rm = "TRUE")

 ## Creating a csv file containing critical dates data
 write.csv2(GRA.med.sdd1, file = paste("GRA median standardised ", Sys.Date(), ".csv", sep=""), row.names = FALSE)

[Package CAVIAR version 2.10-0 Index]